+++ plugins/scaffolder-backend/src/scaffolder/tasks/helper.ts
@@ -15,6 +15,7 @@
  */
 
 import { isArray } from 'lodash';
+import { Schema } from 'jsonschema';
 
 /**
  * Returns true if the input is not `false`, `undefined`, `null`, `""`, `0`, or
@@ -23,4 +24,32 @@ import { isArray } from 'lodash';
  */
 export function isTruthy(value: any): boolean {
   return isArray(value) ? value.length > 0 : !!value;
+}
+
+export function generateExampleOutput(schema: Schema): unknown {
+  const { examples } = schema as { examples?: unknown };
+  if (examples && Array.isArray(examples)) {
+    return examples[0];
+  }
+  if (schema.type === 'object') {
+    return Object.fromEntries(
+      Object.entries(schema.properties ?? {}).map(([key, value]) => [
+        key,
+        generateExampleOutput(value),
+      ]),
+    );
+  } else if (schema.type === 'array') {
+    const [firstSchema] = [schema.items]?.flat();
+    if (firstSchema) {
+      return [generateExampleOutput(firstSchema)];
+    }
+    return [];
+  } else if (schema.type === 'string') {
+    return '<example>';
+  } else if (schema.type === 'number') {
+    return 0;
+  } else if (schema.type === 'boolean') {
+    return false;
+  }
+  return '<unknown>';
 }