+++ packages/backend-test-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/backend-test-utils
 
+## 0.1.25-next.1
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/cli@0.17.2-next.1
+  - @backstage/backend-common@0.13.6-next.1
+
 ## 0.1.25-next.0
 
 ### Patch Changes