+++ plugins/catalog-backend-module-github/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-catalog-backend-module-github
 
+## 0.1.4-next.1
+
+### Patch Changes
+
+- 8335a6f6f3: Adds an edit URL to the GitHub Teams Group entities.
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/backend-tasks@0.3.2-next.1
+  - @backstage/backend-common@0.13.6-next.1
+  - @backstage/integration@1.2.1-next.1
+  - @backstage/plugin-catalog-backend@1.2.0-next.1
+  - @backstage/catalog-model@1.0.3-next.0
+
 ## 0.1.4-next.0
 
 ### Patch Changes