+++ plugins/catalog-graph/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-graph
 
+## 0.2.18-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.1
+  - @backstage/catalog-client@1.0.3-next.0
+  - @backstage/core-plugin-api@1.0.3-next.0
+  - @backstage/catalog-model@1.0.3-next.0
+  - @backstage/plugin-catalog-react@1.1.1-next.1
+
 ## 0.2.18-next.0
 
 ### Patch Changes