+++ plugins/kubernetes/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-kubernetes
 
+## 0.6.6-next.1
+
+### Patch Changes
+
+- b9b8bbc7d9: show request/limit CPU and Memory on the UI
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.1
+  - @backstage/core-plugin-api@1.0.3-next.0
+  - @backstage/catalog-model@1.0.3-next.0
+  - @backstage/plugin-catalog-react@1.1.1-next.1
+  - @backstage/plugin-kubernetes-common@0.3.0-next.1
+
 ## 0.6.6-next.0
 
 ### Patch Changes